prestashop_microservices/requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Flask==2.3.3
Flask-JWT-Extended==4.5.2
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.7
WTForms==3.0.1
Limiter==0.3.1
Flask-Limiter== 3.5.0
Flask-Migrate== 4.0.4
